Improve compatibility of esds boxes
Firefox and Chrome aren't currently able to read varints with a size
different than 4. ISO 14496-1 isn't clear on the matter, although it
limits the maximum varint size to 4.

This patch forces varints to have a size of 4, allowing a broader
compatibility of esds boxes. Sizes greater than 4 are not supported
anymore (but this is dictated by the specification).